Factor out form into separate element
